keymaps changes

+++ src/buffer.c
@@ -117,12 +117,13 @@ enum RemoveResult remove_from_line(struct Document *doc, int y, int x) {
   return REMOVE_LINE;
 }
 
+// TODO: если bx == first.len, last будет равен следующей строке
 void remove_range(struct Document *doc, struct Vec4 c) {
   get_selected_coordinates(&c);
   if (c.ay >= doc->len || c.by >= doc->len) return;
   struct Line *first = doc->buf[c.ay], *last = doc->buf[c.by];
   if (c.ax > first->len || c.bx > last->len) return;
-  int remainder = MAX(last->len - c.bx - 1, 0);
+  int remainder = MAX((long long)last->len - c.bx - 1, 0);
   size_t len = c.ax + remainder;
   if (first->size - 1 < len) {
     first->size = len + ADDITIONAL_REALLOCATION;
